fix `render_section_visibility_compute()` prototype causing camo not to render properly

+++ xlive/Blam/Engine/render/render_section_visibility.cpp
@@ -4,11 +4,11 @@
 void __cdecl render_section_visibility_compute(
 	int32 a1,
 	int32 a2,
-	uint8 model_group_index,
-	int32 a4,
-	int32 a5,
-	int32 a6,
-	int32 a7,
+	int32 model_group_index,
+	uint32 a4,
+	uint32 a5,
+	uint32 a6,
+	uint32 a7,
 	bool a8)
 {
 	INVOKE(0x1A3083, 0x0, render_section_visibility_compute, a1, a2, model_group_index, a4, a5, a6, a7, a8);
